# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditRecord": {
        "additionalProperties": false,
        "description": "Comprehensive, verifiable audit report for a processed email.",
        "properties": {
          "classification": {
            "$ref": "#/components/schemas/ClassificationResult"
          },
          "discrepancies": {
            "items": {
              "$ref": "#/components/schemas/Discrepancy"
            },
            "title": "Discrepancies",
            "type": "array"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentReference"
            },
            "title": "Documents",
            "type": "array"
          },
          "email": {
            "$ref": "#/components/schemas/EmailRecord"
          },
          "email_id": {
            "minLength": 1,
            "title": "Email Id",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/FieldEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "extractions": {
            "items": {
              "$ref": "#/components/schemas/DocumentExtraction"
            },
            "title": "Extractions",
            "type": "array"
          },
          "mismatch_detected": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mismatch Detected"
          },
          "outcome": {
            "anyOf": [
              {
                "enum": [
                  "MATCH",
                  "MISMATCH",
                  "NOT_APPLICABLE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "parsers": {
            "items": {
              "$ref": "#/components/schemas/ParserResult"
            },
            "title": "Parsers",
            "type": "array"
          },
          "partial_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "previous_revision": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Revision"
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingMetadata"
          },
          "result_summary": {
            "minLength": 1,
            "title": "Result Summary",
            "type": "string"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewCase"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "schema_version": {
            "const": "1.0-draft",
            "default": "1.0-draft",
            "title": "Schema Version",
            "type": "string"
          },
          "state": {
            "enum": [
              "COMPLETE",
              "NEEDS_REVIEW"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "email_id",
          "revision",
          "email",
          "classification",
          "state",
          "outcome",
          "mismatch_detected",
          "result_summary",
          "documents",
          "parsers",
          "extractions",
          "evidence",
          "partial_result",
          "discrepancies",
          "review",
          "processing"
        ],
        "title": "AuditRecord",
        "type": "object"
      },
      "ClassificationResult": {
        "additionalProperties": false,
        "description": "Validated intent classification result or proposed review state.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailCategory"
              },
              {
                "enum": [
                  "document_comparison",
                  "new_shipping_instruction",
                  "invoice_query",
                  "general",
                  "spam"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "confidence_indicator": {
            "anyOf": [
              {
                "enum": [
                  "HIGH",
                  "MEDIUM",
                  "LOW"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Indicator"
          },
          "evidence_ids": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "enum": [
              "RESOLVED",
              "NEEDS_REVIEW"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "category",
          "reason",
          "evidence_ids"
        ],
        "title": "ClassificationResult",
        "type": "object"
      },
      "Discrepancy": {
        "additionalProperties": false,
        "description": "A confirmed difference between SI and BL for a comparison field.",
        "properties": {
          "bl_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Bl Value"
          },
          "field": {
            "enum": [
              "shipper",
              "consignee",
              "notify_party",
              "port_of_loading",
              "port_of_discharge",
              "container_count",
              "gross_weight_kg"
            ],
            "title": "Field",
            "type": "string"
          },
          "si_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Si Value"
          }
        },
        "required": [
          "field",
          "si_value",
          "bl_value"
        ],
        "title": "Discrepancy",
        "type": "object"
      },
      "DocumentExtraction": {
        "additionalProperties": false,
        "description": "Seven-field extraction mapping for a single identified document.",
        "properties": {
          "document_id": {
            "minLength": 1,
            "title": "Document Id",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ExtractedField"
            },
            "propertyNames": {
              "enum": [
                "shipper",
                "consignee",
                "notify_party",
                "port_of_loading",
                "port_of_discharge",
                "container_count",
                "gross_weight_kg"
              ]
            },
            "title": "Fields",
            "type": "object"
          },
          "role": {
            "enum": [
              "SI",
              "BL"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "role",
          "fields"
        ],
        "title": "DocumentExtraction",
        "type": "object"
      },
      "DocumentReference": {
        "additionalProperties": false,
        "description": "Document identity, observed role, and identification evidence linkage.",
        "properties": {
          "document_id": {
            "minLength": 1,
            "title": "Document Id",
            "type": "string"
          },
          "identification_evidence_ids": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Identification Evidence Ids",
            "type": "array"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentRole"
              },
              {
                "enum": [
                  "SI",
                  "BL",
                  "UNKNOWN"
                ],
                "type": "string"
              }
            ],
            "title": "Role"
          }
        },
        "required": [
          "document_id",
          "role",
          "identification_evidence_ids"
        ],
        "title": "DocumentReference",
        "type": "object"
      },
...
      "DynamicVerifyRequest": {
        "additionalProperties": false,
        "description": "On-demand verification payload for ad-hoc SI and BL text.\n\nAuthoritative contract citation: 03_DATA_CONTRACTS.md \u00a73.1, line 437.",
        "properties": {
          "bl_text": {
            "title": "Bl Text",
            "type": "string"
          },
          "email_id": {
            "default": "example-on-demand",
            "minLength": 1,
            "title": "Email Id",
            "type": "string"
          },
          "si_text": {
            "title": "Si Text",
            "type": "string"
          }
        },
        "required": [
          "si_text",
          "bl_text"
        ],
        "title": "DynamicVerifyRequest",
        "type": "object"
      },
...
      "EmailRecord": {
        "additionalProperties": false,
        "description": "Ingested email message context and attachment references.",
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentReference"
            },
            "title": "Attachments",
            "type": "array"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "email_id": {
            "minLength": 1,
            "title": "Email Id",
            "type": "string"
          },
          "sender": {
            "title": "Sender",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "email_id",
          "sender",
          "subject",
          "body",
          "attachments"
        ],
        "title": "EmailRecord",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Structured error payload for failed requests or blocked exports (DC-08).",
        "properties": {
          "blocking_emails": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocking Emails"
          },
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "type": "string"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewCase": {
        "additionalProperties": false,
        "description": "Structured human review case containing one or more issues.",
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ReviewIssue"
            },
            "minItems": 1,
            "title": "Issues",
            "type": "array"
          },
          "review_id": {
            "minLength": 1,
            "title": "Review Id",
            "type": "string"
          },
          "state": {
            "enum": [
              "OPEN",
              "RESOLVED"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "review_id",
          "state",
          "issues"
        ],
        "title": "ReviewCase",
        "type": "object"
      },
...
      "ReviewUpdate": {
        "additionalProperties": false,
        "description": "Human review mutation payload. Forbids direct mismatch/outcome mutation (REG-012).",
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewAction"
              },
              {
                "enum": [
                  "CONFIRM",
                  "CORRECT"
                ],
                "type": "string"
              }
            ],
            "title": "Action"
          },
          "actor_id": {
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "added_evidence": {
            "items": {
              "$ref": "#/components/schemas/FieldEvidence"
            },
            "title": "Added Evidence",
            "type": "array"
          },
          "category": {
            "anyOf": [
              {
                "enum": [
                  "document_comparison",
                  "new_shipping_instruction",
                  "invoice_query",
                  "general",
                  "spam"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "classification_evidence_ids": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Classification Evidence Ids",
            "type": "array"
          },
          "corrections": {
            "items": {
              "$ref": "#/components/schemas/FieldCorrection"
            },
            "title": "Corrections",
            "type": "array"
          },
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "review_id": {
            "minLength": 1,
            "title": "Review Id",
            "type": "string"
          },
          "role_corrections": {
            "items": {
              "$ref": "#/components/schemas/RoleCorrection"
            },
            "title": "Role Corrections",
            "type": "array"
          }
        },
        "required": [
          "review_id",
          "expected_revision",
          "actor_id",
          "action",
          "rationale"
        ],
        "title": "ReviewUpdate",
        "type": "object"
      },
...
      "SubmissionRecord": {
        "additionalProperties": false,
        "description": "External evaluation format strictly matching official schema (DC-08).",
        "properties": {
          "category": {
            "enum": [
              "BL_COMPARISON",
              "SI_REQUEST",
              "INVOICE_QUERY",
              "GENERAL",
              "SPAM"
            ],
            "title": "Category",
            "type": "string"
          },
          "defect_fields": {
            "items": {
              "enum": [
                "shipper",
                "consignee",
                "notify_party",
                "port_of_loading",
                "port_of_discharge",
                "container_count",
                "gross_weight_kg"
              ],
              "type": "string"
            },
            "title": "Defect Fields",
            "type": "array"
          },
          "has_defect": {
            "title": "Has Defect",
            "type": "boolean"
          },
          "review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Reason"
          },
          "status": {
            "enum": [
              "OK",
              "MISMATCH",
              "NEEDS_REVIEW"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "category",
          "status",
          "has_defect",
          "defect_fields"
        ],
        "title": "SubmissionRecord",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Shipping document verification, deterministic comparison, evidence tracing, and human review API.",
    "title": "Shipping Document Verification & Audit API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/audit": {
      "get": {
        "description": "Query audit records with optional filters (API-AUD-001).",
        "operationId": "list_audits_audit_get",
        "parameters": [
          {
            "description": "Filter by state (COMPLETE, NEEDS_REVIEW)",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by state (COMPLETE, NEEDS_REVIEW)",
              "title": "State"
            }
          },
          {
            "description": "Filter by category",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category",
              "title": "Category"
            }
          },
          {
            "description": "Filter by outcome (MATCH, MISMATCH, NOT_APPLICABLE)",
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by outcome (MATCH, MISMATCH, NOT_APPLICABLE)",
              "title": "Outcome"
            }
          },
          {
            "description": "Filter by presence of discrepancies",
            "in": "query",
            "name": "has_discrepancy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by presence of discrepancies",
              "title": "Has Discrepancy"
            }
          },
          {
            "description": "Filter only records with confirmed discrepancies",
            "in": "query",
            "name": "mismatch_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "Filter only records with confirmed discrepancies",
              "title": "Mismatch Only",
              "type": "boolean"
            }
          },
          {
            "description": "Filter only records needing human review",
            "in": "query",
            "name": "hitl_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "Filter only records needing human review",
              "title": "Hitl Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditRecord"
                  },
                  "title": "Response List Audits Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audits"
      }
    },
    "/audit/{email_id}": {
      "get": {
        "description": "Retrieve full detail for a single AuditRecord (API-AUD-002, API-AUD-003).",
        "operationId": "get_single_audit_audit__email_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Single Audit"
      }
    },
    "/audit/{email_id}/review": {
      "post": {
        "description": "Submit human review update with optimistic concurrency control (API-REV-001/002/003, DC-07).",
        "operationId": "submit_human_review_audit__email_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Submit Human Review"
      }
    },
...
    "/submission": {
      "get": {
        "description": "T13 HTTP translation only: store snapshot -> ExportService -> adapter.",
        "operationId": "get_submission_submission_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/SubmissionRecord"
                  },
                  "title": "Response Get Submission Submission Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Get Submission"
      }
    },
    "/verify": {
      "post": {
        "deprecated": true,
        "description": "Dynamically audit an SI vs draft BL in real-time (deprecated compatibility shim for T12-04).",
        "operationId": "verify_dynamic_text_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DynamicVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Dynamic Text"
      }
    }
  }
}